move loops to own file
